Change balance to float

+++ qa327/models.py
@@ -21,7 +21,7 @@ class User(db.Model):
     email = db.Column(db.String(100), unique=True)
     password = db.Column(db.String(100))
     name = db.Column(db.String(1000))
-    balance = db.Column(db.Integer)
+    balance = db.Column(db.Numeric)
 
 class Ticket(db.Model):
     """
